fix: 리뷰 반영

+++ src/main/java/com/example/demo/domain/personalartworkcommunication/application/result/PersonalArtworkFeelingReplyLikeResult.java
@@ -5,6 +5,6 @@ import java.time.LocalDateTime;
 public record PersonalArtworkFeelingReplyLikeResult(
     Long personalFeelingReplyId,
     Boolean liked,
-    Integer likeCount,
+    Long likeCount,
     LocalDateTime createdAt,
     LocalDateTime deletedAt) {}
